solidblocks-rds-postgresql-ansible/requirements.txt
ansible==10.7.0
pytest-testinfra==10.2.2
psycopg==3.2.9
